requirements.txt
aiogram==3.13.1
APScheduler==3.10.4
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.1
PyYAML==6.0.2
python-dotenv==1.0.1
tzdata==2024.2
